backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
umap-learn==0.5.5
python-dotenv==1.0.0
pydantic==2.5.0
APScheduler==3.10.4
pytest==7.4.3
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.2
